packages/video-generator/requirements.txt
moviepy
openai
python-dotenv
requests
